Created the Chart component
This char component displays the average price of the vehicle since it was produced. All the data is fictional and calculated through various functions, which basically add / subtract 5% from past year's price to the current year price. What I will try to implement next is an offer section where based on the initial price of the vehicle (as well as the fluctuations throughout the years) and the price it is being sold at, we show whether buying the car is a good/bad offer!