Fixed a bug in the AllCars pagination
Due to limitations of the server the collection of cars has to be filtered in-memory... Fixed an issue where if you filter the cars and end up with only 2 offers, the pagination component would still show that there are 6 more pages of offers (considering that the max count of offers per page is 5, that shouldn`t be happening). Now when you refilter the cars, the pagination page is reset to one and all the cars show up properly!